Remove unused imports and unecessary casts, foreachify loops
git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@1625 edc67d0a-ffea-dc11-983e-003048321218

+++ src/main/java/com/threerings/opengl/gui/event/KeyEvent.java
@@ -66,7 +66,7 @@ public class KeyEvent extends InputEvent
     {
         // TEMP: This is a hack to get around a bug in lwjgl's handling of
         // numpad keys in windows
-        if ((int)_keyChar == 0) {
+        if (_keyChar == 0) {
             switch (_keyCode) {
             case Keyboard.KEY_NUMPAD1: return '1';
             case Keyboard.KEY_NUMPAD2: return '2';
